Factor MTMetaworld wrapper into MultiEnvWrapper
- Add flag to MultiEnvWrapper to account for task_one_hot
  that is already appended to metaworld multi task benchmark
  environments
